+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-Version 2.5.0+dev
+Version 2.6.0-dev
 - marked on* properties of FileLoader as deprecated
 - option onUntranslated of Nette DIC extension throws a deprecation warning with non-default values
+- deprecated MessageSelector in favor of new IntervalsMessageSelector
 
 Version 2.5.0
 - added TomlLoader